Create README.md
This program generates a random date by using the Random class to select a random year between 1900 and 2021, a random month between 1 and 12, and a random day between 1 and the number of days in the selected month. The resulting random date is then outputted to the console in the format of "yyyy-MM-dd".